geometry-priors/segmentation_requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu117
torch
torchvision
numpy
pandas
tqdm
argparse
pyarrow
Pillow
transformers
timm
accelerate
kornia